Hot fix

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,8 +9,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int bit_t;
-	int count = 0;
+	unsigned long int bit_t;
+	unsigned long int count = 0;
 
 	bit_t = n ^ m;
